recommendation document

+++ Smartmeet/src/main/java/tn/esprit/examen/Smartmeet/repositories/MaryemSalhi/IFeedbackRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tn.esprit.examen.Smartmeet.entities.MaryemSalhi.Feedback;
 
-public interface IFeedbackRepository extends JpaRepository<Feedback, Integer> {
+public interface IFeedbackRepository extends JpaRepository<Feedback, Long> {
 }
